+++ include/cipher.hpp
@@ -7,7 +7,7 @@
 class cipher
 {
     private:
-        static const int block_size;
+        static const size_t block_size;
     public:
         virtual void encrypt(char *s) = 0;
         virtual void decrypt(char *s) = 0;
